requirements.txt
unidecode
pinyin_jyutping_sentence
pykakasi